Consolidate client-side duplication (issue #68 mechanical items)
- Hoist shared DOM-based esc() into public/shared.js, loaded before
  explore.js and ride.js; both now reference window.esc
- Extract LINE_COLORS into src/data/rail-colors.ts as single source of
  truth; Rail.tsx imports from shared module, app.js gets pointer comment
- Document judgment items: dist() and ETA terminal-branch differences
