+++ public/explore.js
@@ -333,11 +333,7 @@
 
   // ─── Utilities ───
 
-  function esc(str) {
-    const el = document.createElement('span');
-    el.textContent = str;
-    return el.innerHTML;
-  }
+  const esc = window.esc;
 
   function debounce(fn, ms) {
     let timer;
