+++ public/ride.js
@@ -397,11 +397,7 @@
     return R * 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
   }
 
-  function esc(str) {
-    const el = document.createElement('span');
-    el.textContent = str;
-    return el.innerHTML;
-  }
+  const esc = window.esc;
 
   // ─── Boot ───
   if (document.readyState === 'loading') {
